Fix Memory to BaseMemory
